backend/requirements.txt
# Paquetes esenciales
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0

# Base de datos
sqlalchemy==2.0.15
psycopg2-binary==2.9.6  # Requerido para conexión PostgreSQL + SSL

# Autenticación
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Validación de datos
pydantic==1.10.7  # Compatible con FastAPI < 0.100
email-validator==2.2.0

# Generación de códigos QR
qrcode[pil]==7.4.2

# Criptografía para QR seguro
cryptography==41.0.1